Google_PubSub/requirements.txt
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
enum34==1.1.6
futures==3.2.0
google-api-core==0.1.4
google-auth==1.4.0
google-cloud-pubsub==0.30.1
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.9.1
idna==2.6
protobuf==3.5.1
psutil==5.4.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pytz==2018.3
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
